refactor: move RowKeyWrapper from core to hbase module
This class is used before RPC, we need this class for various filter calculation.
 -  None of the class from `bigtable-core-client` referenced `RowKeyWrapper`, So removing it from source location.
 - RowKeyWrapper refers `ByteStringComparator#compare` for moved the compare logic in it.
