+++ bigtable-client-core-parent/bigtable-hbase/src/main/java/com/google/cloud/bigtable/hbase/util/RowKeyWrapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google Inc. All Rights Reserved.
+ * Copyright 2020 Google LLC.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.bigtable.util;
+package com.google.cloud.bigtable.hbase.util;
 
 import com.google.api.core.InternalApi;
 import com.google.protobuf.ByteString;
@@ -34,6 +34,36 @@ public class RowKeyWrapper implements Comparable<RowKeyWrapper> {
 
   @Override
   public int compareTo(RowKeyWrapper o) {
-    return ByteStringComparator.INSTANCE.compare(key, o.key);
+    return compare(key, o.key);
+  }
+
+  private int compare(ByteString key1, ByteString key2) {
+    if (key1 == null) {
+      if (key2 == null) {
+        return 0;
+      } else {
+        return 1;
+      }
+    } else if (key2 == null) {
+      return -1;
+    }
+
+    if (key1 == key2) {
+      return 0;
+    }
+
+    int size = Math.min(key1.size(), key2.size());
+
+    for (int i = 0; i < size; i++) {
+      // compare bytes as unsigned
+      int byte1 = key1.byteAt(i) & 0xff;
+      int byte2 = key2.byteAt(i) & 0xff;
+
+      int comparison = Integer.compare(byte1, byte2);
+      if (comparison != 0) {
+        return comparison;
+      }
+    }
+    return Integer.compare(key1.size(), key2.size());
   }
 }
